Module might be attached to a non-existing ELB.
An ELB resource might not be realised from declaration, i.e count is
set to 0. In that case the module invocation will fail because of the
variables for a non existing resource are accessed. This commit
replaces the output values with empty ones.

For e.g

resource "aws_elb" "myelb" {
  count = "${var.createelb == 0 ? 0 : 1 }"
  name                     = "${var.base_name}-elb"
}

module "ssl-policy-elb" {
  create = "${var.createelb == 0 ? 0 : 1 }"
  source = "$GH/terraform-aws-ssl-ciphers/ELBSecurityPolicy-TLS-1-2-2017-01"
  name = "public-ssl-policy"
  load-balancer-id = "${element(concat(aws_elb.myelb.*.id, list("")), 0)}"
}